backend/requirements.txt
# Django and REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-filter>=23.2
django-cors-headers>=4.0.0

# Database
psycopg2-binary>=2.9.6

# Authentication
djangorestframework-simplejwt>=5.2.2

# API Documentation
drf-yasg>=1.21.5

# Environment Variables
python-decouple>=3.8

# Date and Time
python-dateutil>=2.8.2

# Excel/CSV Processing
openpyxl>=3.1.2
pandas>=2.0.0

# Utilities
pytz>=2023.3
Pillow>=10.0.0

# Development Tools
black>=23.3.0
flake8>=6.0.0
pytest>=7.3.1
pytest-django>=4.5.2

# Production Server
gunicorn>=20.1.0
whitenoise>=6.4.0